Removed the unusued files of declarations
