--- src/clipboard.test-d.ts
@@ -1,4 +1,0 @@
-import { expectType } from 'tsd';
-import Clipboard from './clipboard';
-
-expectType<Clipboard>(new Clipboard('.btn'));
